Refactor and bring up the service with some unit testing
